requirements.txt
fastapi
uvicorn
pandas
scikit-learn
tldextract
python-Levenshtein
joblib
pymongo
tldextract
python-Levenshtein
